feat: add back ethereum-vsl
