+++ internal/engine/worker/decentralized/factory.go
@@ -29,6 +29,7 @@ import (
 	"github.com/rss3-network/node/internal/engine/worker/decentralized/contract/savm"
 	"github.com/rss3-network/node/internal/engine/worker/decentralized/contract/stargate"
 	"github.com/rss3-network/node/internal/engine/worker/decentralized/contract/uniswap"
+	"github.com/rss3-network/node/internal/engine/worker/decentralized/contract/vsl"
 	"github.com/rss3-network/node/internal/engine/worker/decentralized/core"
 	"github.com/rss3-network/node/schema/worker/decentralized"
 )
@@ -77,6 +78,8 @@ func New(config *config.Module, databaseClient database.Client, redisClient ruei
 		return kiwistand.NewWorker(config)
 	case decentralized.SAVM:
 		return savm.NewWorker(config)
+	case decentralized.VSL:
+		return vsl.NewWorker(config)
 	case decentralized.Stargate:
 		return stargate.NewWorker(config)
 	case decentralized.Curve:
